+++ crates/zng-app/src/view_process.rs
@@ -600,11 +600,27 @@ event_args! {
             list.search_all()
         }
     }
+
+    /// Arguments for the [`VIEW_PROCESS_SUSPENDED_EVENT`].
+    pub struct ViewProcessSuspendedArgs {
+
+        ..
+
+        /// Broadcast to all widgets.
+        fn delivery_list(&self, list: &mut UpdateDeliveryList) {
+            list.search_all()
+        }
+    }
 }
 
 event! {
-    /// View Process finished initializing and is now online.
+    /// View-Process finished initializing and is now online.
     pub static VIEW_PROCESS_INITED_EVENT: ViewProcessInitedArgs;
+    /// View-Process suspended, all resources dropped.
+    ///
+    /// The view-process will only be available if the app resumes. On resume [`VIEW_PROCESS_INITED_EVENT`]
+    /// notify a view-process respawn.
+    pub static VIEW_PROCESS_SUSPENDED_EVENT: ViewProcessSuspendedArgs;
 }
 
 /// Information about a successfully opened window.